Update Armor.md
The combat system uses a location chart to determine the armor resistance for a given impact. In many situations, such as a called shot, there is no armor on the intended target location while in the case of a miss or normal attack the area of impact can be armored or not to implement this change we would have to make two changes. First convert resistance tables to reflect material as opposed to armor configuration ie, cloth/padded, leather, wood, copper, bronze, iron, steel, titanium, Kevlar, advanced composites. Second add an armor diagram to track armored locations suggest Front and Back view of body such that the location number is identified and the armor can be shaded or written in to reflect what the character is wearing. 
Added in Slashing damage type as it differs significantly depending on armor construction and material type.
All numbers are based loosely on hardness and durability data from numerous sources.