Ideas and Suggestions
Use this file to field ideas and float suggestions for further consideration. Each section is listed and comments should be sorted into the sections they relate to. This file should never appear in the core rules or world content. As ideas are reviewed the contributions that get into the core will be recorded in the appropriate tomb and the reference here will be moved or removed Ideas that have been deemed inappropriate for inclusion will be recieve a strike through and be moved to the bottom of the file.