Corrected logical errors and conflicts
Removed "The base Oddities follow the *Disposition System*, which outlines six major points of Disposition. The *Disposition System* is not exclusive to characters. It can be used to characterize any group of people (cities, countries, kingdoms, non-profit organizations, etc.) to create Oddities that influence the group." so that it can be added to Game System expansion section